fixed sql id problem on saving a category added remove and a test to best both functions

+++ src/main/java/be/smal/games/repositories/GameJPARepository.java
@@ -3,5 +3,5 @@ package be.smal.games.repositories;
 import be.smal.games.domain.Game;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GameJPARepository extends JpaRepository<Game, Long> {
+public interface GameJPARepository extends JpaRepository<Game, Integer> {
 }
